requirements.txt
# Core dependencies
python>=3.10
fire>=0.5.0
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# LLM clients
openai>=1.12.0
anthropic>=0.18.0
tiktoken>=0.5.0

# MCP (Model Context Protocol)
mcp>=0.9.0

# Async and networking
httpx>=0.26.0
aiohttp>=3.9.0

# Data processing
pydantic>=2.5.0
tenacity>=8.2.3

# Logging and monitoring
rich>=13.7.0
zmq>=0.0.0
pyzmq>=25.1.0

# Utilities
pyyaml>=6.0.1
markitdown>=0.0.1a2

# Optional: Browser automation (if using playwright tools)
playwright>=1.40.0

# Optional: Search and web scraping
firecrawl-py>=0.0.1
exa-py>=1.0.0

